[refactor]: co-locate causal Wan and denoising stages
