+++ fastvideo/pipelines/basic/turbodiffusion/turbodiffusion_pipeline.py
@@ -7,12 +7,13 @@ for 1-4 step video generation using rCM (recurrent Consistency Model) sampling
 with SLA (Sparse-Linear Attention).
 """
 
+from fastvideo.pipelines.basic.wan.stages.denoising import WanDenoisingStage
 from fastvideo.fastvideo_args import FastVideoArgs
 from fastvideo.logger import init_logger
 from fastvideo.models.schedulers.scheduling_rcm import RCMScheduler
 from fastvideo.pipelines import ComposedPipelineBase, LoRAPipeline
-from fastvideo.pipelines.stages import (ConditioningStage, DecodingStage, DenoisingStage, InputValidationStage,
-                                        LatentPreparationStage, TextEncodingStage, TimestepPreparationStage)
+from fastvideo.pipelines.stages import (ConditioningStage, DecodingStage, InputValidationStage, LatentPreparationStage,
+                                        TextEncodingStage, TimestepPreparationStage)
 
 logger = init_logger(__name__)
 
@@ -52,11 +53,11 @@ class TurboDiffusionPipeline(LoRAPipeline, ComposedPipelineBase):
                                                     transformer=self.get_module("transformer", None)))
 
         self.add_stage(stage_name="denoising_stage",
-                       stage=DenoisingStage(transformer=self.get_module("transformer"),
-                                            transformer_2=self.get_module("transformer_2", None),
-                                            scheduler=self.get_module("scheduler"),
-                                            vae=self.get_module("vae"),
-                                            pipeline=self))
+                       stage=WanDenoisingStage(transformer=self.get_module("transformer"),
+                                               transformer_2=self.get_module("transformer_2", None),
+                                               scheduler=self.get_module("scheduler"),
+                                               vae=self.get_module("vae"),
+                                               pipeline=self))
 
         self.add_stage(stage_name="decoding_stage", stage=DecodingStage(vae=self.get_module("vae"), pipeline=self))
