+++ fastvideo/pipelines/basic/wan/stages/__init__.py
@@ -1,0 +1,2 @@
+# SPDX-License-Identifier: Apache-2.0
+"""Wan sampling and conditioning. Import each stage from its defining module."""
