+++ fastvideo/pipelines/pipeline_batch_info.py
@@ -132,6 +132,8 @@ class ForwardBatch:
     raw_latent_shape: tuple[int, ...] | None = None
     noise_pred: torch.Tensor | None = None
     image_latent: torch.Tensor | None = None
+    # Normalized clean first frame for Wan TI2V and causal-DMD conditioning.
+    first_frame_latent: torch.Tensor | None = None
 
     # Action control inputs (Matrix-Game)
     mouse_cond: torch.Tensor | None = None  # Shape: (B, T, 2)
