+++ fastvideo/tests/stages/_denoising_fixtures.py
@@ -111,6 +111,7 @@ def _patch_denoising_module(monkeypatch, cfg_gate_step):
         expected_gate_step = float(cfg_gate_step)
 
     import fastvideo.pipelines.stages.denoising as denoising
+    from fastvideo.pipelines.stages.base import PipelineStage
 
     # envs.py evaluates FASTVIDEO_CFG_GATE_STEP lazily via __getattr__, so the
     # stage sees monkeypatched values without reloading the module.
@@ -122,4 +123,56 @@ def _patch_denoising_module(monkeypatch, cfg_gate_step):
     monkeypatch.setattr(denoising, "get_world_group", lambda: SimpleNamespace(local_rank=0))
     monkeypatch.setattr(denoising, "get_attn_backend", lambda **kwargs: object())
     monkeypatch.setattr(denoising, "set_forward_context", lambda **kwargs: nullcontext())
+    monkeypatch.setattr(PipelineStage, "device", property(lambda self: torch.device("cpu")))
     return denoising, logger
+
+
+class RecordingDenoiser(TinyDenoiser):
+
+    def __init__(self, offset=0.0):
+        super().__init__()
+        self.offset = offset
+        self.inputs = []
+
+    def forward(self, hidden_states, prompt_embeds, timestep, guidance=None, encoder_hidden_states_image=None):
+        self.inputs.append((hidden_states.clone(), timestep.clone()))
+        prompt = prompt_embeds[0].float().mean()
+        self.calls.append("uncond" if prompt < 0 else "cond")
+        time = timestep.float().reshape(hidden_states.shape[0], -1).mean(dim=1).reshape(-1, 1, 1, 1, 1)
+        return hidden_states[:, :2].float() * 0.125 + prompt * 0.25 + time * 0.0001 + self.offset
+
+
+class TinyVAE(torch.nn.Module):
+
+    def __init__(self):
+        super().__init__()
+        self.weight = torch.nn.Parameter(torch.zeros(()))
+        self.shift_factor = torch.tensor([0.25, -0.25]).reshape(1, 2, 1, 1, 1)
+        self.scaling_factor = torch.tensor([2.0, 0.5]).reshape(1, 2, 1, 1, 1)
+        self.encode_calls = 0
+
+    def encode(self, image):
+        self.encode_calls += 1
+        return SimpleNamespace(mean=torch.full((1, 2, 1, 2, 4), 0.5, dtype=torch.float32))
+
+
+def _batch(steps=4, cfg=True):
+    batch = _tiny_batch()
+    batch.latents = torch.linspace(-0.5, 0.5, 48).reshape(1, 2, 3, 2, 4)
+    batch.num_inference_steps = steps
+    batch.timesteps = torch.linspace(1000, 1, steps)
+    batch.do_classifier_free_guidance = cfg
+    batch.num_frames, batch.height, batch.width = 9, 16, 32
+    batch.raw_latent_shape = tuple(batch.latents.shape)
+    batch.return_trajectory_latents = True
+    return batch
+
+
+def _args():
+    args = _tiny_args()
+    args.vae_cpu_offload = True
+    arch = SimpleNamespace(patch_size=(1, 2, 2))
+    args.pipeline_config.dit_config.arch_config = arch
+    args.pipeline_config.vae_config = SimpleNamespace(
+        arch_config=SimpleNamespace(scale_factor_temporal=4, scale_factor_spatial=8))
+    return args
